decouples RecordWriter from Record
