+++ RecordWriter.py
@@ -1,13 +1,13 @@
+# Python Imports
 from typing import Optional, Dict
 import re
-import inspect
 
+# GitEHR Imports
 from helper_functions import get_current_datetime
 
-
-class RecordWriter:
+class Record:
     """
-    Returns a markdown-formatted string, to be written to a GitEHR Record file.
+    Class to create a GitEHR markdown record including YAML, contents, and PGP signing.
     """
 
     def __init__(
@@ -17,30 +17,6 @@ class RecordWriter:
         self.separator = separator
         self.meta_data = meta_data
 
-        self.add_yaml_meta(self.meta_data)
-
-    def add_yaml_meta(self, meta_data) -> None:
-        """Adds initial meta data to frontmatter contents."""
-        self._add_initial_frontmatter(meta_data)
-
-    def _add_initial_frontmatter(self, data: dict) -> None:
-        """Adds frontmatter to top of contents str."""
-
-        frontmatter_str = self._create_frontmatter_str(data)
-
-        # Appends frontmatter to beginning of contents str
-        self.contents = "\n".join([frontmatter_str, self.contents])
-
-    def _create_frontmatter_str(self, data: dict) -> str:
-        """Method to create a YAML frontmatter string including the key-values provided."""
-
-        initial_frontmatter = ["---", "---"]
-
-        for key, val in data.items():
-            initial_frontmatter.insert(-1, f"{key}:{val}")
-
-        return "\n".join(initial_frontmatter)
-
     def add_line(self, content: str) -> None:
         self.contents += content + self.separator
 
@@ -48,6 +24,11 @@ class RecordWriter:
         _contents_lst[0] = "\n" + _contents_lst[0]
         _contents_joined = "\n".join(_contents_lst)
         self.add_line(_contents_joined)
+
+    def _prepend_yaml(self) -> None:
+        yaml = YAMLFrontmatter(self.meta_data)
+        yaml_string = yaml.get_string()
+        self.contents = "\n".join([yaml_string, self.contents])
 
     def _add_public_key(self) -> None:
         self.add_contents(
@@ -59,24 +40,48 @@ class RecordWriter:
             ]
         )
 
-    def get_contents(self, sign=False) -> str:
+    def get_contents(self, sign=True) -> str:
+        self._prepend_yaml()
+
         if sign:
             self._add_public_key()
+
         return self.contents
 
+class RecordWriter:
+    """Takes a Record object and writes to file."""
+    
+    def __init__(self, record_obj:Record, file_extension:str='.md'):
+        self.file_extension=file_extension
+        self.record = record_obj
+    
+    def write(self, filename)->None:
+        """Takes Record object's contents and writes to PATH: {destination_path}/{filename}"""
+
+        with open(filename, "w") as entry_file:
+            
+            contents=self.record.get_contents()
+
+            entry_file.write(contents)
+        
 
 class YAMLFrontmatter:
     """Helper Class to aid interaction with YAML.
-    
+
     Instantiate with a dictionary of meta data items.
-    
+
     Useful methods:
-        1) `.add_yaml_items()` -> takes a dictionary of meta data items to add.
-        2) `.get_string()` -> turns current contents into a YAML string
-        3) `.extract_yaml_from_string()` -> searches input string and returns YAML contents, or None if one can't be found.
+        - `.add_yaml_items()` -> takes a dictionary of meta data items to add.
+        - `.get_string()` -> turns current contents into a YAML string
+        - `.extract_yaml_from_string()` -> searches input string and returns YAML contents, or None if one can't be found.
+        - `.get_meta_data()` -> getter
     """
+
     def __init__(self, meta_data: dict = None):
         self.meta_data = meta_data
+
+    def get_meta_data(self) -> dict:
+        return self.meta_data
 
     def _convert_meta_dict_to_list(self, meta_dict: dict) -> list[str]:
         return [f"{key}:{val}" for key, val in meta_dict.items()]
@@ -87,9 +92,9 @@ class YAMLFrontmatter:
     def extract_yaml_from_string(self, input_string: str) -> str:
         matches = re.search(r"---\n(([\s|\S])*)(?<!-)---(?=\n)", input_string)
         if matches:
-            return matches.group(0).replace('\n    ','\n')
-    
-    def add_yaml_items(self, items_to_add:dict)->None:
+            return matches.group(0).replace("\n    ", "\n")
+
+    def add_yaml_items(self, items_to_add: dict) -> None:
         if self.meta_data is not None:
             self.meta_data.update(items_to_add)
         else:
@@ -104,25 +109,9 @@ class YAMLFrontmatter:
 
 
 if __name__ == "__main__":
-    yaml_obj = YAMLFrontmatter(
-        {
-            "created_on": get_current_datetime(),
-            "created_by": "PLACEHOLDER",
-        }
+    new_entry = Record(
+        "This is a new entry for Patient A.\nHe presented with dyspnoea.\nManagement is xyz.",
+        meta_data={"created_on": get_current_datetime(), "created_by": "Dr AC"},
     )
-    inpt_str = """---
-    created_on:2023-07-18
-    created_by:PLACEHOLDER
-    ---
-    Hi guys
-    This is an entry for Patient
-    Patient presented today
-    Management is...
-    -----BEGIN PGP PUBLIC KEY BLOCK-----
-    mQINBFRUAGoBEACuk6ze2V2pZtScf1Ul25N2CX19AeL7sVYwnyrTYuWdG2FmJx4x
-    =nUop
-    -----END PGP PUBLIC KEY BLOCK-----"""
 
-    yaml = yaml_obj.extract_yaml_from_string(inpt_str)
-
-    print(yaml)
+    print(new_entry.get_contents())
